Refactor how I like it
