+++ dispatcher.yml
@@ -8,8 +8,7 @@ service:
     scaledown_wait: 15  # seconds
   main_kwargs:
     node_id: demo-server-a
-  metrics_kwargs:
-    log_level: debug
+  metrics_kwargs: {}
 worker:
   worker_kwargs:
     idle_timeout: 3
